msvc C4293

+++ src/algo/hash/_hash_.c
@@ -7,5 +7,5 @@ void
 print_hash(const char *name, const char *ss, unsigned long hash)
 {
   printf("%s\n------------\n%s -> %16lu\n", name, ss, hash);
-  printf(BPRI64(hash));
+  printf(BPRI64((uint64_t)hash));
 }
